requirements.txt

Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==11.0.0
django-storages==1.14.2
boto3==1.34.69
django-filter==24.1
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
drf-spectacular==0.27.2
gunicorn==22.0.0
whitenoise==6.7.0
